requirements.txt
fastapi
selenium
python_dotenv
fints
